door monitor (#28)
* initial ESP8266 with WiFi

* add door switch sensor

* update time with NTP

might be wise to timestamp our door events at the source

* add MQTT publish

* only publish state change to mqtt

* change door states to booleans

* add timestamp to mqtt msg

should be more accurate than relying on subscribers to set the time (especially if we make these sticky)

* retain last mqtt message

* initialize door state to invalid value

if we publish only changed values, there is a chance we won't see the initial bootup value if it is the same as the initialized value. So initialize to something that is sure to change.