SCD40 updates (#37)
* remove the no-serial blocker

* SCD4x add last updated time property

Need to track both the last time the sensor was read, and the last time it was successfully read. Otherwise on read error we'll keep trying to read it every loop, which is likely too fast for the sensor to recover.

* add eCO2 to environmental display

* Convert pressure to in Hg

* don't MQTT publish old SCD40 data

SCD40 sometimes stops providing data. If this happens we don't want to keep publishing the last value every MQTT interval, especially since we don't publish a timestamp.

* attempt to re-start the SCD40 on read failure

this fails to solve the dying SCD40 problem

* add pressure correction to SCD40

* don't mqtt publish too-stale sensor readings

similar to what we did for SCD40 previously